This submission addresses Trac tickets 15, 21, 22, 24, 25, & 26.
This corrects several of the inconveniences of building CS-MAP on Linux 64 bit systems.  Also note, that the perennial prolem of _wcsicmp has been resolved by writing our own equivalent: CS_wcsicmp in CS_ansi.c

Also note that the compilation failure on 64 bit gcc was due to the fact that a "long" in 64 bit gcc (and also Sun, by the way) is a 64 bit object.  CS-MAP uses the "long32_t" and "ulong32_t" wherever it is expecting a 32 bit integer value.  This useage should/must be adhered to in order to avoid similar problems in the future.

git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@1555 4e78687f-474d-0410-85f9-8d5e500ac6b2

+++ Source/CS_mgrs.c
@@ -630,8 +630,8 @@ int CScalcMgrsFromLlUtm (struct cs_Mgrs_ *__This,char *result,int size,double la
 		*rsltPtr++ = setPtr [setIdx].easting [idxX];
 		*rsltPtr++ = setPtr [setIdx].northing [idxY];
 	}
-	sprintf (utmXascii,"%05ld",(intUtmX % 100000));
-	sprintf (utmYascii,"%05ld",(intUtmY % 100000));
+	sprintf (utmXascii,"%05d",(int)(intUtmX % 100000));
+	sprintf (utmYascii,"%05d",(int)(intUtmY % 100000));
 	for (ii = 0;ii < prec;ii += 1)
 	{
 		*rsltPtr++ = utmXascii [ii];
